refactor(v0.2): create core/reasoning/ package skeleton (#29 phase 1/3)
Phase 1 of 3 for the reasoning split. Pure structural move — no
behavior change. Establishes the namespace for two follow-up PRs
that will extract `query()` body into `modes.py` (mode dispatch)
and `pipeline.py` (Loop body).

Move (git mv, history preserved):
- core/reasoning_engine.py → core/reasoning/engine.py (1040 LOC, 51 KB)

Public API (core/reasoning/__init__.py) re-exports the 3 symbols
call sites currently import from core.reasoning_engine:
- ReasoningEngine
- MAX_LOOP
- LOOP_TIMEOUT

Module-level constants not used externally (CONFIDENCE_TH,
TIMING_TARGET_SEC, SYSTEM_LOG_PATH) remain accessible only via
deep imports from `core.reasoning.engine`.

23 import statements rewritten across 6 entry points:
- core/graph_rag_engine.py: 1 (the wrapper)
- james_phase6_gate.py: 8
- james_phase6_test.py: 5
- james_phase55_test.py: 4
- james_diagnostic.py: 4
- james_phase5_test.py: 1

Plus 3 dynamic imports (`import core.reasoning_engine as re_mod`
inside `inspect.getsource()` checks) updated to deep import
`core.reasoning.engine` so `inspect` still finds the source.

Verification:
- python -c "from core.reasoning import ReasoningEngine, MAX_LOOP,
  LOOP_TIMEOUT" clean
- python -m compileall -q core/ clean
- ReasoningEngine.__module__ == "core.reasoning.engine"
- STEP 7 12-query benchmark (cp949 default console — exercises
  PR #36 fix too): 11/12 OK + 1 BLOCK + 1 ERR (q12 timeout, same
  pre-existing slow-query flake as baseline). Total 402.2s vs
  baseline 380.9s (Δ +5.6%, within LLM-noise band documented in #35).
- Invariants preserved: q11 security block byte-identical (26
  bytes, 0.0s) across all 4 runs to date. q1/q3/q5/q7/q10/q11
  graph_paths IDENTICAL across 4 runs — pure-refactor confirmed
  at the structural level.

Out of scope (#29 follow-ups):
- engine.py is still 51 KB, exceeds the 20 KB gate from #29.
  Phase 2 will extract chat/wiki/self_evolve/coding/web_fallback
  mode dispatch from query() body into modes.py.
- Phase 3 will extract Loop 0/1/2 body into pipeline.py.
