+++ core/reasoning/engine.py
