+++ james_phase6_gate.py
@@ -69,7 +69,7 @@ def run_e2e_stability():
 
     def t_empty_context_fallback():
         """빈 컨텍스트 → '자료에 없음' 즉시 반환 (LLM 호출 없음)"""
-        from core.reasoning_engine import ReasoningEngine
+        from core.reasoning import ReasoningEngine
         engine = ReasoningEngine()
         answer = engine._generate_answer("테스트", "")
         ok = "자료에 없음" in answer and len(answer) > 0
@@ -77,7 +77,7 @@ def run_e2e_stability():
 
     def t_short_context_fallback():
         """50자 미만 컨텍스트 → fallback"""
-        from core.reasoning_engine import ReasoningEngine
+        from core.reasoning import ReasoningEngine
         engine = ReasoningEngine()
         answer = engine._generate_answer("테스트", "짧음")
         ok = "자료에 없음" in answer
@@ -85,7 +85,7 @@ def run_e2e_stability():
 
     def t_llm_error_fallback():
         """LLM 에러 prefix → 의미 있는 fallback 반환"""
-        from core.reasoning_engine import ReasoningEngine
+        from core.reasoning import ReasoningEngine
         engine = ReasoningEngine()
         # _LLM_ERROR_PREFIXES 존재 확인
         has_prefix = hasattr(engine, "_LLM_ERROR_PREFIXES")
@@ -95,7 +95,7 @@ def run_e2e_stability():
 
     def t_blocked_result_structure():
         """보안 차단 결과 구조 완전성"""
-        from core.reasoning_engine import ReasoningEngine
+        from core.reasoning import ReasoningEngine
         result = ReasoningEngine._blocked_result("보안 차단 테스트")
         required = {"answer","blocked","graph_paths","graph_used","sources","timing_sec"}
         has_all  = all(k in result for k in required)
@@ -113,13 +113,13 @@ def run_e2e_stability():
 
     def t_loop_injection_defense():
         """MAX_LOOP=2 고정 — loop 주입 불가"""
-        from core.reasoning_engine import MAX_LOOP
+        from core.reasoning import MAX_LOOP
         ok = MAX_LOOP == 2
         return ok, f"MAX_LOOP={MAX_LOOP} (고정값=2 필수)"
 
     def t_loop_timeout_exists():
         """LOOP_TIMEOUT 설정 존재"""
-        from core.reasoning_engine import LOOP_TIMEOUT
+        from core.reasoning import LOOP_TIMEOUT
         ok = 0 < LOOP_TIMEOUT <= 60
         return ok, f"LOOP_TIMEOUT={LOOP_TIMEOUT}s"
 
@@ -471,8 +471,8 @@ def run_performance():
 
     def t_loop_count_enforced():
         """Loop 횟수 MAX_LOOP=2 코드 레벨 강제"""
-        from core.reasoning_engine import MAX_LOOP
-        import inspect, core.reasoning_engine as re_mod
+        from core.reasoning import MAX_LOOP
+        import inspect, core.reasoning.engine as re_mod
         src  = inspect.getsource(re_mod)
         # range(MAX_LOOP + 1) 패턴으로 3번까지 (0,1,2)
         uses_max_loop = "range(MAX_LOOP + 1)" in src or "range(MAX_LOOP+1)" in src
@@ -481,7 +481,7 @@ def run_performance():
 
     def t_context_size_limited():
         """Context 크기 제한 — 800자 이상 LLM에 전달 안 됨"""
-        from core.reasoning_engine import ReasoningEngine
+        from core.reasoning import ReasoningEngine
         import inspect
         src = inspect.getsource(ReasoningEngine._generate_answer)
         has_limit = "800" in src or "[:800]" in src
@@ -518,7 +518,7 @@ def run_performance():
 
     def t_latency_target_code_exists():
         """latency 목표값 코드 존재 확인"""
-        import inspect, core.reasoning_engine as rm
+        import inspect, core.reasoning.engine as rm
         src = inspect.getsource(rm)
         has_target = "TIMING_TARGET_SEC" in src or "30" in src
         return has_target, f"latency 목표 코드 존재={has_target}"
